Criação de autenticação do usuário e validação no mesmo para acesso as rotas

+++ server/src/routes/memories.ts
@@ -3,14 +3,22 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
-    const memories = prisma.memory.findMany({
+  app.addHook("preHandler", async (request) => {
+    // Aqui é utilizado o Jwt para fazer a verificação do token e assim realizar a utilização da rota.
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
+    const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createAt: "asc",
       },
     });
 
-    return (await memories).map((memory) => {
+    return memories.map((memory) => {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
@@ -19,7 +27,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
       };
     });
   });
-  app.get("/memories/:id", async (request) => {
+
+  app.get("/memories/:id", async (request, reply) => {
     // Aqui utilizo o Zod para validar o meu ID sendo uma string, e assim sendo mais fácil de trabalhar com ele no request.params
     const paramsSchema = z.object({
       id: z.string().uuid(),
@@ -34,8 +43,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     });
 
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
     return memory;
   });
+
   app.post("/memories", async (request) => {
     const bodySchema = z.object({
       content: z.string(),
@@ -50,27 +64,39 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "2f797770-5cec-46f6-b62a-22692d85b9f3",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
     });
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
+    
     const { id } = paramsSchema.parse(request.params);
 
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -84,12 +110,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories:id", async (request) => {
+  app.delete("/memories:id", async (request, reply) => {
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
     const { id } = paramsSchema.parse(request.params);
 
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
     await prisma.memory.delete({
       where: {
         id,
